insurance_ai/requirements.txt
flask
flask-cors
werkzeug

pdfplumber
python-docx
pytesseract
Pillow

torch
transformers
sentence-transformers
faiss-cpu

numpy
scikit-learn
